Make badgr_server_slug optional for BadgeClass object
badgr_server_slug field was added to BadgeClass by cherry-picking
commit 3309aab2a2eb00d28c5ca3d3145c8dddb15e6159 but it is causing some erroneous
behaviour in the LMS. That field should not be mandatory since the BadgeClass is
used for badging in general and not neccessarily with the Badgr backend.

Even when using the Badgr API the BadgeClass object will be created in the LMS
before we request to create a corresponding item in Badgr and the badgr_server_slug
value itself is retrieved from Badgr after successfully creating that corresponding
item.

(cherry-picked from 2c47dc9696ed32482441f6f07c41cc65637e1301)